Move PathsList component to its own folder
