+++ static/src/components/Paths/PathsList.jsx
